Main

+++ database/migrations/2024_01_01_000005_create_answers_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->foreignId('assessment_id')->constrained()->onDelete('cascade');
             $table->foreignId('question_id')->constrained()->onDelete('cascade');
             $table->text('answer_text')->nullable();
-            $table->integer('score')->nullable();
+           $table->decimal('score', 5, 2)->nullable();
             $table->string('attachment_path')->nullable();
             $table->timestamps();
         });
